requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
Django==4.2
django-channels==0.7.0
django-cors-headers==4.3.1
django-storages==1.14.2
djangorestframework==3.14.0
drf-yasg==1.21.7
idna==3.6
inflection==0.5.1
oauthlib==3.2.2
packaging==23.2
pillow==10.2.0
psycopg2-binary==2.9.9
python-decouple==3.8
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.10.0
uritemplate==4.1.1
urllib3==2.2.1
